+++ src/detect-transform-urldecode.c
@@ -86,7 +86,9 @@ static bool BufferUrlDecode(const uint8_t *input, const uint32_t input_len, uint
             if (i + 2 < input_len) {
                 if ((isxdigit(input[i+1])) && (isxdigit(input[i+2]))) {
                     // Decode %HH encoding.
-                    *oi = (input[i+1] >= 'A' ? ((input[i+1] & 0xdf) - 'A') + 10 : (input[i+1] - '0')) << 4;
+                    *oi = (uint8_t)((input[i + 1] >= 'A' ? ((input[i + 1] & 0xdf) - 'A') + 10
+                                                         : (input[i + 1] - '0'))
+                                    << 4);
                     *oi |= (input[i+2] >= 'A' ? ((input[i+2] & 0xdf) - 'A') + 10 : (input[i+2] - '0'));
                     oi++;
                     // one more increment before looping
